Refactor model schemas
