+++ adaptive_hockey_federation/tests/model_schemas/generics.py
@@ -1,21 +1,22 @@
 from tests.model_schemas.fields_validation_schemas import (
-    CORRECT_CITY_NAMES,
+    ALL_LETTERS,
     CORRECT_CREATE,
     CORRECT_UPDATE,
     FIGURES_AND_LETTERS,
     FIGURES_ONLY,
     LONG_256,
     LONGER_THEN_256,
-    NOT_CYR,
     NULL,
+    PUNCTUATION_MARKS_ONLY,
+    THE_ONLY_LETTER,
 )
 
-CITY_MODEL_TEST_SCHEMA = {
+SIMPLE_UNIQUE_NAME_MODEL_TEST_SCHEMA = {
     CORRECT_CREATE: {
-        "name": "Городкоторогонет",
+        "name": "Какоетоимя",
     },
     CORRECT_UPDATE: {
-        "name": "Новыйгородкоторогонет",
+        "name": "Какоетоновоеимя",
     },
     "must_not_be_admitted": (
         {
@@ -23,15 +24,20 @@ CITY_MODEL_TEST_SCHEMA = {
             "test_values": (
                 LONGER_THEN_256,
                 NULL,
-                NOT_CYR,
                 FIGURES_ONLY,
+                PUNCTUATION_MARKS_ONLY,
+                THE_ONLY_LETTER,
             ),
         },
     ),
     "must_be_admitted": (
         {
             "fields": "name",
-            "test_values": (LONG_256, FIGURES_AND_LETTERS, CORRECT_CITY_NAMES),
+            "test_values": (
+                LONG_256,
+                FIGURES_AND_LETTERS,
+                ALL_LETTERS,
+            ),
         },
     ),
 }
