+++ adaptive_hockey_federation/tests/model_schemas/nosology.py
@@ -1,0 +1,3 @@
+from tests.model_schemas.generics import SIMPLE_UNIQUE_NAME_MODEL_TEST_SCHEMA
+
+NOSOLOGY_MODEL_TEST_SCHEMA = SIMPLE_UNIQUE_NAME_MODEL_TEST_SCHEMA
